driver/onload: fix spinlock unlocking in error path
Fix error path in handover code, which handles the case when the file is
changing under Onload's feet.  It looks like the error path was not
updated when locking schema in all the oo_fd_replace_file() function was
changed, and file_lock spin lock was replaced by task_lock.

Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>