ON-15599: fix bogus assertion
The test should be that the packet index is in the range [0,superbuf_pkts)
as it was before my recent changes. Comparison with "end" is not valid for
"prev" since the off-by-one fiasco means they sometimes refer to different
superbufs.

Also moved the assertion to catch any attempt to use a bogus pkt_id to
access packet (meta)data.