ON-15590: Set ipif scope at creation time
It is based on the user's patch to complete the initialisation
of the new ipif entry before use. Otherwise, cplane makes the
wrong cp_ipif_notify_oof() invocations based on the entry's
last scope.

At startup, it made spurious invocations of cp_ipif_notify_oof()
for the host-scoped addresses, such as localhost.

In runtime, if the recycled entry was the host-scoped in the past,
the cplane mistakenly skipped the cp_ipif_notify_oof() invocation
regardless of its new scope.