ON-15590: Do not add filters for host-scoped addresses
This is a bugfix that prevents Onload from attempting to set up
filters for the host-scoped listening sockets, which caused the
following error:

$ sudo ip addr add 10.0.0.1/24 scope host dev enp4s0f0
$ onload nc -l 10.0.0.1 1234
$ dmesg
[onload] __oof_socket_add: 2:2047 ERROR: laddr=10.0.0.1 not local

This patch makes such arbitrary host-scoped addresses similar to
the localhost (which is also host-scoped).

The correct behaviour is achieved by filtering out the wrong ipif
entries in the cicp_user_addr_is_local_efab() with the new "compare
scope" checker.