Cleanup: Separate evq number and rxq number
Currently it is assumed that the rxq instance number will match the evq
it is bound to. This will need to change for ef10ct nics which we are
planning to use evqs that have multiple rxqs bound to them. Equivalent
code already existed for tx, so this is mostly just doing the same
thing.