ON-15222: Bind to ef10ct LL rx queues in efhw
In order to test ef10ct LL rx in ef_vi/onload we need to be able to bind
to an rxq. Due to the differences between efct and ef10ct we will have
to manage the queues ourselves.
There are also some changes to the test driver to match the
implementation in efhw.