ON-15241: Add generic software buffer table implementation to efhw
On previous architectures: the nic contained the buffer table
(ef10,ef100), we didn't use a buffer table (efct), or we had a software
implementation (afxdp). For ef10ct we will want to use the buffer table
to store the pages which will be used as superbufs and to translate
between the physical address and dma addresses when not in physical
addressing mode.

This commit separates out the existing software buffer table
implementation from afxdp into a more generic form (called efhw_sw_bt)
which can be used by multiple architectures (currently afxdp and ef10ct
in the next commit).

Due to the similarities between the "umem" functions and the existing
functionality provided by iobufset, I have moved iobufset into a more
generic location (lib/kernel_utils) where it can be used by both efthrm
and efhw. iobufset files are basically unchanged in order to keep this
patch simpler, though this means that there are still some functions
with the "oo_" prefix.