ON-15240: Post rx buffers to ef10ct nic directly
In cases where the user application can use physical addressing mode
then it would be preferable to post superbufs from user space rather
than paying the cost of a user->kernel transition. This commit adds a
new io region that can be mapped into userspace when using physical
addressing mode, this is then used to post the superbufs without the
cost of an ioctl.