ON-15577: Split hwports in netif into TX/RX hwports
Replace hwport_mask with tx_hwport_mask and rx_hwport_mask in the
ci_netif and ci_netif_state data structures.

Before multiarch NIC architecture, Onload allocated resources for both
TX and RX per net_device in the accelerated Onload stacks. The patched
hwport_mask was sufficient in that case.

The multiarch NIC offers two datapaths (FF and LL) per physical port,
represented as two hwports (FF and LL) per net_device. The Onload users
must choose which of these they want to use for the accelerated stacks,
specifically for TX and RX. They might select FF for TX and both FF and
LL for RX. This way, the existing hwport_mask cannot express this
configuration, so this patch suggests splitting it into the TX and RX
parts.

Functionally, this is a no-operation patch, but it paves the way for the
upcoming unidirectional datapath support.

Suggested-by: Siân James <sian.james@amd.com>