ON-15583: Identify LL and FF multiarch NIC datapaths
Compute llct_hwports and ff_hwports and apply them to the previously
introduced rx_hwport_mask and tx_hwport_mask. Currently, we disable yet
unsupported LLCT RX, but this should change as soon as we introduce new
or repurpose the existing EF_ knobs to select multiarch NIC datapaths.

It is straightforward to identify the LLCT datapaths thanks to the
specific NIC_FLAG_LLCT flag.

The FF datapaths are trickier because, currently, they look identical to
the "plain" SFC adapters. This might or might not change in the future.
Meanwhile, to distinguish, this patch uses the fact that LLCT and FF
datapaths of the multiarch NICs refer to the same net_device object.