ON-15583: Handle zero-sized RXQs
This patch uses the previously introduced rx_hwport_mask to identify
ef_vi instances that will not be used for RX. For such ef_vi instances,
Onload will allocate zero-sized RXQs.

While ef_vi already supports zero-sized RXQs, it required us to make a
few Onload workflows conditional:

* ci_netif_init_fill_rx_rings()
* ci_netif_rx_post()
* ci_netif_set_rxq_limit()
* ef_vi_receive_set_discards()

This patch does not address the OOF manager due to its limited API.
When Onload registers a new socket and installs a new filter, it cannot
submit the port mask and tell the OOF manager where to accept or drop
packets. By default, OOF installs filters on all announced and
brought-up ports. As a result, when a packet arrives, but there is a
zero-sized RXQ, it is discarded, which is visible with ethtool:

$ ethtool -S enp4s0f0 | grep port_rx_dp_q_disabled_packets