ON-15692: Use x64_sys_call function when syscall_table isn't available
On previous kernel versions `do_syscall_64` used to call the specific
syscall implementation indirectly via a syscall table. On newer kernels
this has been replaced by a call to a new function `x64_sys_call` which
takes `struct ptregs` and the syscall number as arguments.

This change updates the code that parses the compiled kernel to look for
the syscall table to also look for this new function as well. Since it
is expected that finding either of these will be mutually exclusive the
searching function returns after finding one.

The syscall dispatch code for x86_64 has been modified to handle either
having the syscall table or the function.

Tested on old versions (still using syscall table):
 * RHEL7.9
 * RHEL9.3

Tested on new kernels:
 * Debian12 with kernel version 6.1.0-20
 * Ubuntu 22.04 with kernel version 5.15.156