ON-15561: Fix test_multi_ns*[ipvlan] cplane sysunit tests
The affected tests failed because they did not specify the IPVLAN
mode. The IPVLAN mode selected by default, L3 (1), is not accelerated
by Onload, that only supports L2 (0). As a result, when the test case
tried to resolve a test route, the resulting HW port mask was empty,
i.e. unexpectedly unaccelerated.

To fix it, the test case passes additional options for using the L2
mode when it creates an IPVLAN interface.