ON-15658: support both locations of EFCT metadata
Introduce a 'meta_offset' field to 'ef_vi_efct_rxq_ptr' so that
'data_pkt' can be correctly updated using simple arithmetic on the
critical polling path.

In practice it must be zero or one, otherwise more complex logic
would be needed to handle more than one dislocated packet after
rollover. It has the opposite value to the associated design
parameter (zero for "off-by-one" and one for "in the same buffer"),
so I didn't attempt to make it look like an enumeration of the design
parameter.