ON-15742: match the new CONFIG_RETPOLINE option name for v6.9+ kernels
The CONFIG_RETPOLINE option was renamed to CONFIG_MITIGATION_RETPOLINE
which meant that Onload would incorrectly assume the option was not
enabled. This resulted in failure to find the syscall table on kernels
with this option enabled. This first impacted v6.9-rc1 and was added
in torvalds/linux@aefb2f2e619b6c334bcb31de830aa00ba0b11129 [^1].

[^1]: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=aefb2f2e619b6c334bcb31de830aa00ba0b11129