ON-15749: Define EFX_NOT_UPSTREAM in onload build
Ideally we wouldn't have to know about this, and the bug will remain open to
explore better solutions. This change allows us to avoid having to re-patch
the net driver on every import until we get a long term fix, which will become
even more painful with the auxbus support.