cleanup: link unit tests with -ldl
It can be useful to intercept library functions to reduce dependencies
that are difficult to satisfy in a unit test (e.g. the test for
lib/ciul/efct_ubufs intercepts mmap to remove the dependency huge pages).
We want to allow dlsym to be used to access the original function, so
link with the library which provides that.