epoll3: fix socket cleanup after duplicating epoll fd
After closing a dup()'ed epoll fd all the sockets which were added
to that fd get invalid epoll reference in citp_fdinfo.
In epoll3 case it leads to a refcounting issue - when the socket
is closed we do not release reference to netif taken in
citp_epoll_set_home_stack().

Keep list of all opened epoll fds associated with this epoll set
and set valid epoll_fd to all sockets referenced to the epoll fd
which is being closed.

Signed-off-by: Sergey Nikitin <sergey.nikitin@oktet.tech>