ON-15821: Add support for increased ci_resource_alloc_t sizes
For new versions of the char driver to maintain backwards compatibility
with ef_vi application compiled against older versions we have to ensure
that user/kernel interface doesn't change. However with the addition of
rx_post_buffer_mmap_bytes field any new changes would increase the size
of ci_resource_alloc_t, and so when returning from kernel to user it
could overwrite adjacent bytes causing corruption.

This change repurposes the existing out_id field to now be in a union
with a new in_struct_size field which is used by the user application to
inform the char driver the expected sizeof ci_resource_alloc_t. To
maintain backwards compatibility the value of the field is added to 88
(the current size in bytes). This means that old application which
memset the struct to 0 will work as before, and the char driver won't
try to update these fields and cause corruption.

Any future field will have to be if-guarded to check whether the
endoffset fits within the input struct size.