ON-15486: Fix rollover logic in test driver
Previously the test driver would check for rollovers after writing a
packet. This works fine when there are more suberbufs available to be
used, however when no superbufs are available the test driver still
thinks it has written a packet and decrements the remaining packet
counter which could result in receiving fewer packets than expected.

This commit fixes this by moving the check for rollover before packet
writing and finishes the rx loop there so that the packet is written on
the next iteration (if a superbuf is available).