ON-15829: relax constraints on new EFCT design parameters
When exposing a new design parameter to ef_vi, we conservatively prevent old
apps (linked against an ef_vi which does not know about the new parameter)
from using NICs with a non-default parameter value (on the grounds that the
app might be assuming the default value and might go wrong if the value is
different).

That causes sprurious failure in cases where there are already multiple values
in use and old apps do not make assumptions. Specifically, there are X3
variants with 8 and 16 queues, which caused failures when exposing DP_RX_QUEUES
even though ef_vi doesn't currently care about the value.

This change allows us to relax the constraint in cases where we know it's safe.