ON-15851: Re-enable objtool for onload modules
Previously disabled as part of ON-7109 due to excessive warnings during
compilation. Since then the output of objtool is used to live-patch
thunks when loading the kernel modules, and as of linux 6.9 it will
print a stack trace if the unpatched return thunk is used. Re-enabling
the objtool prevents this issue at module load time, but also
re-introduces the build time warnings. The objtool warning all relate to
the use of noreturn functions which are not known to the upstream
kernel.