ON-15856: assert that TX timestamps can store subnano bits on init
cbc293192c (ON-15865: populate ts_flags and ts_nsec_frac for efct_vi
TX events, 2024-06-11) added sub-nanosecond bits to TX timestamps on
EFCT VIs. This also included an assertion that enough space existed
for storing these given the run-time parameter ts_subnano_bits. But
such code may exist on a hot path, so this commit moves this to the
more reasonable location of after this design parameter is grabbed.