lib/unix: fix fdtable allocation if its size is too large
If EF_FDTABLE_SIZE is not set, RLIMIT_NOFILE hard limit is used as
fdtable size. The value may be too big. For instance, on Debian 13
it is 1073741816=3ffffff8. In such case allocation will fail.
Introduce a new environment variable EF_FDS_MAX to limit maximum
size of fdtable. By default it is 1048576 (default value of RLIMIT_NOFILE
hard limit on Debian/Ubuntu).

Signed-off-by: Sergey Nikitin <sergey.nikitin@oktet.tech>