ON-14517: Improve handling of filter insertion failures during nic reset
Before this patch if an operation that led to filter insertion (e.g.
bind, IP_MEMBERSHIP_ADD, listen) took place at the same time as a nic
was resetting then there was a chance that a filter could be installed
on a nic while onload thought that no filter was installed. This issue
is especially easy to provoke on a bond over multiple physical nics.

Commit 4bcea86 added some functionality for handling this scenario but
only implemented it for `oo_socket_steal_or_add_wild`. This patch
expands on this previous work to cover more scenarios.