ON-15863: Factor out common efct filter code.
Much of the efct filtering will be common between efct and ef10ct. This is a
non-functional change to make the filtering code agnostic to the underlying
hardware.

As more of the ef10ct stuff gets implemented there are likely to be more changes
but this is intended to get the bulk of the changes out of the way up front to
minimise conflicts.