ON-15289: Fix cpu2rxq path in sfcaffinity script
A path to the cpu2rxq file has changed after Onload transitioned
from using the driverlink interface to the AUX bus in c055714c83.

For instance, a path:
  /sys/class/net/enp4s0f0/device/sfc_resource/cpu2rxq

Has become:
  /sys/class/net/enp4s0f0/device/sfc.onload.0/sfc_resource/cpu2rxq

The new path contains the subdirectory "sfc.onload.0", named after
the corresponding AUX device name, which this patch must search
with RE or glob because the AUX device suffix is unknown in advance.