ON-14517: Stop using vestigial reset handling code
efrm_driverlink_desist and efrm_driverlink_resume are functions which
were added to stop us using driverlink operations while the nic was
down. However, the desist function was only called when filter removal
failed with -ENETDOWN on an ef10 nic. In reality this mean that it was
encountered very infrequently (and never on other nic arches).

I haven't fully removed the definitions as they are still used in the
driverlink code, so this can be removed along with the rest of the
driverlink code.