ON-15945: Install citools static library when installing package
If this is installed along with the rest of onload then we can avoid
rebuilding onload source when building tcpdirect.

Also include the generated onload_version.h in the package contents so
that when we are building tcpdirect from installed onload we have a
record of the exact version that was used to build it.