ON-15885: Remove unix server module
unix_server.h was only used by the shrub server and became quite coupled
with it. For example, the epoll_data for a socket determines which
callback is used and this data is determined by the shrub_server. Its
methods have been moved into shrub_server.c because of this.