ON-15712: Increase offset to 64 bits to prevent overflow
Previously, the `offset` field of `oo_pkt_buf_map_t` was only 32 bits.
This meant that if a stack allocated > 4GiB of memory from the hugetlb,
when the userland part of onload tried to map in the memory for packet
buffers which had a larger offset than 4GiB the offset parameter would
rollover and it would map in bad data. This patch increases the width of
the `offset` field to 64-bits which should fix the issue until we have
128 bit machines and someone wants > 18.45 exabytes of packet buffers.