ON-15862: Check for presence of auxbus in vi_alloc/free
The vi_allocator for efct and ef10ct is stored their arch_extra which
has the same lifetime as auxbus. It is therefore only safe to
allocate/free when auxbus is present.

Though communication with the driver through auxbus isn't needed for
these allocations, the alternative would be to add another lock to
ensure the vi_allocator isn't freed while we alloc/free a vi.