ON-15862: ef10: Prevent use after free of arch_extra
The vi_allocator in ef10's arch_extra is freed during driver unload and
so we must check if it is still there. As with x3, check if auxbus is
still present to gate access to arch_extra. To allow this, move arch_extra
construction/destruction to ef10_probe/ef10_remove respectively.