eflatency: compute the mean from the timings array
This patch changes how the mean column in eflatency is computed
to use the sum of the per-message timings, instead of the
overall time that the test used.  That way things that are outside
of the timing loop do not affect the computation of the mean and
the mean value reflects the min/50/90/95/max values in the other
columns.

For most users this will result in a slight decrease of their mean
times, although ideally it should be very close for the default case.