ON-15852: Adjust CI_GLOBAL_WORKQUEUE to use an unbound workqueue
This was after experimental results running MONP in combination with wq-monitor
with periodic dumps to see how often the CPUitsv counter increments.

The alternatives here was to use the old mechanism to be bound to the same core
that allocs the workqueue. This is problematic given this core is busy and results
in the CPUitsv counter incrementing.

The current solution enables the use WQ_UNBOUND, this was deemed safe despite the
potential risk of work item reordering. This is deemed safe because work items
are protected here via oo_filter_ns refcounts.