ON-15920: Store hwport<->ifindex relationship in cp shim
The cplane sysunit tests can occasionally fail if the cp server is
unable to store the relationship between hwport and ifindex. I suspect
that this is caused by the cp server processing the updates "out of
order", specifically processing the new hwport message before the
netlink RTM_NEWLINK message. With the normal cp_server when it has
finished handling the new link it will request a dump of hwport from the
onload kernel module, which should fix any issues with processing the
messages out of order. However, with the shim cp server used for the
sysunit tests when this hwport dump is requested it receives no updates
as the cp shim doesn't keep track of this information which will leave
the cp server in an incorrect state which eventually leads to the test
failing.

This patch makes the shim cp keep track of the hwport announcements by
storing it in a table which is used when receiving a
OO_IOC_CP_DUMP_HWPORTS request from the cp server after it has processed
a new link. With this change the cp server can recover from a previous
failure and the tests can pass.

So that the table can be used by both the cp server and the sysunit
tests it is put at the end of the CP_SHIM_FILE (which acts a proxy for
/dev/onload) after the MIB and FWD bytes.