ON-14137: Allow time to drain outstanding events on stack shutdown.
As part of the ul helper migration the limit was changed from using jiffies,
but the replacement code neglected the shift needed to compare values between
khz and frc, resulting in often only having time for a single iteration of the
loop. Because ci_netif_poll is limited by max events (64 by default) this meant
that any time we had more than 64 events outstanding we ended up emitting the
warning.