ON-14221: Log tx error counter when warning about missing TX completions
We know that we get this error triggered by CTPIO re-ordering resulting in
TX error events, but it requires additional checking to inspect the log to
find the error event being reported. Logging the stat here makes it easier
to check that we are only seeing this in association with TX errors.