ON-16092: Add base_headers option to onload_install
This option includes the header files that were previously installed by
default (onload extensions and ef_vi headers). This option is on by
default to maintain backwards compatability with tools that use
onload_install. However, debian packages won't install if there is a
file that has been installed by another package, so this option is
disabled when creating the onload-dev debian package.