ON-16076: Document behaviour of EF_SHARED_RXQ_NUM with bad queue.
There are various conditions where the requested queue may not be available. These
range from fixed condtions, such as a queue out of range of the hardware
capabilities of the NIC, to dynamic conditions, such as the requested queue being
in use elsewhere and unshareable, or fewer queues than the NIC can support being
created due to hardware capabilities of the host. In theory we could detect the
static problem, but in practice it's the dynamic limits that are likely to be more
problematic, so just document the behaviour.