ON-15889: use efct_vi free list to manage local buffers
A free buffer list was added to efct_vi at some point to work around
problems with the X3 user-kernel interface. For X4 we can use this to
manage the free pool of buffers with Onload sharing the pool between
its user and kernel interfaces.

This avoids having to share a shrub buffer pool with the kernel instance.