ON-16118: Extract efct specific bits from common efrm_efct_rxq handling
On X3 we use this resource type to manage attachment to the underlying HW RXQs.
Much of this management is common with what we want to do for ef10ct, including
tracking attachments and managing wakeups. The queue type managed currently
contains various efct specific information. This change pulls that out, to
allow use of the common resource management on both efct and ef10ct arches.