ON-16156: Add func op to efhw_nic to get location of rxq window
This function gives the io address of the rxq window for a specified
rxq. This will allow for mapping the individual rxq windows into
userspace, rather than having to calculate the offset into the entire
region.

I have left the existing superbuf_io_region op unchanged, this is done
to ensure that this commit series is bisectable but I plan to remove it
once the main body of this ticket is completed.