ON-16156: Add support for mapping rxq windows to the rxq resource type
Extend the memory mapping functionality added in a previous commit, by
adding an implementation for EFCH_VI_MMAP_RX_BUFFER_POST. This allows
for mapping the window for a specific rxq rather than the entire window.

Also, update efrm_rxq_alloc so that it gets and stores the io address of
the window after successfully binding a shared rxq. This stored value is
used when the memory mapping later takes place.