ON-16156: Remove unused superbuf mapping code
Most of the removed code would map the entire rxq window region, so was
no longer compatible with mapping each window individually. The rest of
the removed code was the support for mapping the rxq window region in
the vi resource type, this is now handled as part of rxq resource type
and so was removed.