ON-15996: Add version field to efx_auxdev in test driver
Add this to align the test driver's aux abi with sfc-linux-net's. Check
this field against the driver version contained in efx_auxbus.h which
onload is compiled with. On major version mismatch or provided minor
less than onload's, exit ef10ct_probe with error to prevent use
of an incompatible api.