ON-16171: Get filter types supported by ef10ct via mcdi.
This commit updates ef10ct so that rather than reporting a hardcoded set
of filter types, it insteads issues a GET_PARSER_DISP_INFO MCDI command
to get the supported filter types for the LL datapath.

Commit also includes changes to the test driver to read the MCDI command
an fill out a response.