ON-15635: Move arch -> page map type mapping into efhw
Currently the decision on how to map pages for use by the hardware is done
based on a combination of declared arch and inferred from presence of a pci
dev. This commit makes each arch declare the mapping type it requires and
updates the queue mapping and buffer mapping code to use that information.