ON-16119: Support five-tuple filters for efct nics in efhw
Update definition and uses of struct ef_hw_filter to include remote ip
and port information. This allows for tracking of which five-tuple
filters are supported in hardware.

Also, pass in the nic's set of filter_flags so that we can determine
which parts of the hw filters to compare when checking for equality.
This allows for a distinction between efct and ef10ct which have
slightly different sets of support filters.