Cleanup: Move sanitise_ethtool_flow into efct.c
sanitise_ethtool_flow is a function used to update the request for a hw
filter into a version that is acceptable by the X3's filter support.
Since this function only has relevance for a single nic architecture it
should be moved out of the generic filter handling file and into the
nic-specific file.