ON-16207: Remove explicit handling for exclusive ownership of rxq 0
Existing code assumes that rxq 0 is a special case and is handled by the
netdriver and thus isn't suitable for exclusive ownership. This is true
for X3 nics where packets will default to queue 0, but for X4 LL
datapath there are no special considerations for any queues.

This commit removes the explicit handling of queue 0, meaning that for
X4 LL it can be exclusively used like any other queue. To ensure that
existing X3 code doesn't break, a dummy filter entry is updated stating
it uses queue 0 and that it has non-exclusive ownership.