ON-16227: resolve build issues with C++ compilers
Some C++ apps, such as TCPDirect, consume the ef_vi API headers
directly. As such, some workarounds are needed to ensure those
compilers are happy with the definitions. This patch does two
things to satisfy these compilers.

The first change is to use the namespaced definition of ef_vi ops in
the ef_vi compatibility data struct. The second of which is to cast
the value of `vi->nic_type.arch` to `enum ef_vi_arch` in the function
`ef_vi_get_real_arch`. The second solution avoids the need to change
the underlying type of `arch` in the `ef_vi_nic_type` struct from
from `unsigned char` to `enum ef_vi_arch` which would increase the
size from one byte to four bytes.