ON-15889: expose EFCT buffer descriptor structures
These are currently used more widely than efct_vi.c to manage a free list
of buffers in shared state, requiring a side interface to access them.
We'll want to do more complicated things with this to support ef10ct buffer
management, so it makes sense to expose the structure and reduce the side
interface to a simple descriptor access function.