ON-16224: Release vi reference when unbinding from LL rxq
Previously, if multiple efct_rxq resources had bound to the same LL rxq,
then only the last one to unbind would release it reference to the
underlying vi and pd resources. This meant that if a vi attach to an
existing rxq, when it unbound it would leak the vi resource, then at
driver unload time it would fail an assert.

This commit, update the ef10ct unbind function so that the freer
callback is still called if there are still users on the queue.