ON-15889: manage EFCT buffer queue in shared state
We already had the ability to put the buffers into a linked list, in order
to manage the free buffers. This adds the ability to manage the set of
buffers that have been posted to the NIC and are waiting to be consumed by
the VI.

The state is slightly more complex than the free list: it's a FIFO rather than
a stack; we need to track both those which have been filled by the NIC (and
therefore can be replaced in the NIC FIFO) and those which have been passed
to the VI; and we need to track how many buffers are in the NIC FIFO to
enforce the limit.

This allows us to share exclusive queue management between user and kernel
instances of an Onload VI. Shared queue management is out of scope for now.