ON-16160: Keep track of the current state on an ef10ct rxq.
In order to handle concurrent access and updates to an rxq we must
ensure that we only bind/unbind to it when it is in the correct state.
Most updates to the rxq (alloc, free, init) are performed with
synchronous fw_rpc operations, so these are easy to track within efhw,
however the fini operation is asynchronous. This can potentially lead to
situations where a simple ref count is insufficient to guard access to
the queue.

For example, when the last user of an rxq unbinds efhw will submit a
fini requeust to firmware. In this case, if a new user tries to bind to
it before the flush event has been received it will see that the current
ref count is 0, and then attach to it and re-init the rxq. Then, when
the previous flush event has been received and is processed, efhw will
free the underlying rxq while the new user still thinks it can use it.