ON-16160: Test driver: fini rxq when freed.
Refactor rxq teardown logic from the existing efct_test_free_rxq function
into a new efct_test_fini_rxq function. This new function is called for
both rxq fini and rxq free. Previously, there were situations where an
rxq would be used after it was freed, but since the rxq free function
had no effect on the initialisation of the queue this wouldn't lead to
any errors. Now if a queue is used after it was freed it will fail.