Cleanup: Factor rxq resource allocation into a function in ubufs
Previously attaching to an rxq as a shrub client would lead to UBSAN
warnings as the n_hugepages parameter wasn't filled. A similar issue was
fixed for non-shared queues in 5664ca1, but this didn't cover shared
queues. This commit makes both local and shared attach functions use the
same code for the resource allocation as they currently all use the same
values, but updates/fixes on one codepath may not propagate to the
other.