ON-16027: fix ef_vi compat timestamp assertions
Before merging the original commit set to add the ef_vi compat layer,
it was suggested that the assertions created an all-ones value by
taking the bitwise not of zero instead of casting negative one to an
unsigned integer of appropriate width. My testing was insufficient to
catch a quirky C feature where values shorter than sizeof(int) are
promoted to integers when evaluating a bitwise not operation. In this
case, ~(uint16_t)0 is equivalent to ~(int)(uint16_t)0 which means
that the result of bitwise right-shifts were implementation-defined,
and typically would be sign-extended. As a result, the right-shift
would still yield all-ones while the assertion wanted it to be
zero-extended. To illustrate:
```
(~(uint16_t)0) >> 2
= (~(int)0u) >> 2    (type promotion)
= (~0) >> 2
= 0xffffffff >> 2
= 0xffffffff         (sign-extended right shift)
```

The new approach just directly writes out the hex masks to make it
easier and clearer what is going on.