ON-15705: always use ef_vi arch with ef_vi_init
There are a few places in the onload code base that call ef_vi_init,
one of which (in tcp_helper_resource) would directly pass the EFHW
arch of the NIC instead of the ef_vi arch. While these are supposed
to be the same, ef_vi doesn't actually use EF_VI_ARCH_EF10CT in
favour of EF_VI_ARCH_EFCT. This meant that Onload VIs initialised
in tcp_helper_resource wouldn't be fully initialised, although this
would be missed as we ignore the return value of ef_vi_init.