ON-16264: Pass evq kernel virt address to enable evq
ef10ct and efct used 'phys_to_virt' on the evq dma_addrs to get the
kernel virtual address of the evq. This is not recommended for use on
physical addresses and not allowed for dma bus addresses. Instead,
pass through the virtual address of the evq backing memory we had
already.