ON-16280: Handle multi-datapath NICs when installing unicast filters in oof
This change does the filter install loop twice - the first time on our
preferred hwport, the second time for a fallback hwport for any of the
perferred ports that failed.

For single datapath NICs there is no fallback port. For multi-datapath NICs
the fallback port is the FF datapath. This will be used if we couldn't
insert a filter on the preferred LL datapath, which could be because
1 the stack in question is not using the LL datapath
2 the filter in question is not supported on the LL datapath
3 the filter cannot be installed on the LL datapath for another reason, such
  as lack of space

The most likely case here is 1, which is also the case that fails earliest,
as the stack will not have a vi on that datapath to attempt installation with.

In practice 2 should only occur for IPv6 filters. A stack that is using
IPv6 would normally expect to be configured to use the FF datapath, so would
usually fall into case 1. However, if they are using both datapaths we will
try installing the filter, which should be detected in the lower levels of SW.
In theory we could use knowledge of what filter is being requested and what
the HW supports to avoid this, but the additional complexity is not worth it.

In the case of 3 we can't detect this without actually asking the HW to
install the filter anyway.