ON-16207: Don't print filter information for dummy hw filters
The debugfs output for efct prints information for the currently
installed filters. Now that we are using a dummy filter to specify the
ownership of rxq 0, we shouldn't output this alongside the real hw
filters. A dummy driv_id for the filter is used to determine if the
filter is real or not. A value of -1 is chosen for this dummy id value
since the X3 net driver uses this value to index in to a table, so it
should give negative/extremely large numbers.