ON-16283: calculate min_nic_order on pd allocation
This value was hard-coded to 0, which was fine before as supported
hardware either had no software buffer table, or had a minimum NIC
order of 0. However, with the addition of ef10ct, this isn't safe
to assume anymore as ef10ct only allows order 9 allocations. This
change finds this value from the list of orders instead of making
the assumption that all hardware that supports buffer tables also
supports order 0 allocations.