ON-16304: Identify acceleratable interfaces better
Onload checks the hwport mask of the llap and fwd entries in the mib
tables to decide whether it can accelerate the stack. We normally
require the strict inclusion of these hwports into the hwport mask
requested by netif.

However, this should not apply to the multiarch nics whose datapaths
might be unselected while the underlying netdev is still selected.
This requires us to "hide" the multiarch nics in the predicates, which
want to disambiguate between multiarch nics and bonds to allow
incomplete inclusion for the former and disallow for the latter.

To achieve that, ci_netif_get_hwport_mask() will set all multiarch
hwports regardless of the datapath configurations to "hide" the nics
multiarch nature.