ON-15884: Remove unnecessary work queue flush on rxq bind
This flush was initially added to ensure that any flush events were
processed before updating the state of the rxq. However, the logic that
this was depending on has been moved to a different point in the unbind
process and is no longer dependent on the work queue processing of the
flush events.