ON-16283: improve RX method reporting in eflatency
Commit 7769d358e75a ("ON-16277: allow tx/rx datapath selection in
eflatency") added functionality to select different RX and TX modes
in eflatency. Part of this included updating the reporting to show
this split, but this could ambiguously report the RX mode on some
NICs (such as X3 which would report "ff"). This commit updates this
reporting to report the type of event returned by ef_vi for an RX
event (i.e., EF_EVENT_TYPE_RX or EF_EVENT_TYPE_RX_REF).