ON-16283: remove X3 reference in eflatency mode reporting
This mode is selected where allocation of a VI happens on an interface
that only supports CTPIO sends. The fact that this interface only
supports CTPIO sends does not really matter to the reporting, as it
just uses CTPIO and is indistinguishable from the ctpio_test mode with
the exception of the test name.