ON-16319: Don't request wakeups on dummy ef10ct evqs
If we request an interrupt it on a dummy evq it will lead to a firmware
failure. Prevent this by returning early if the wakeup is request on a
dummy evq. Kernel only ef10ct evqs should be automatically re-primed, so
nothing else needs to be done in this case.