ON-16319: efrm vi: Only request queues which have non-zero capacity
Previously when allocating a vi from efrm_vi_resource_alloc it would use
the default values (true) for the `want_txq` and `want_rxq` vi
attributes.

Now it will check the capacities for the rxq and txq before doing the
allocation. This will stop us from allocating a hardware evq for an
ef10ct vi with only an rxq and no txq.