ON-16444: Increase minimum RXQ size on X4
We use hugepages for superbufs, so will allocate memory for two buffers.
Using only one buffer is not a sensible thing to do, as it will always leave
a point where the ring is empty between the buffer being filled and reposted.
Increasing the minimum size here means that we'll always use at least two
superbufs.