ON-15884: Interrupt driven flush completion for X3
This is only expected to be relevant for a single txq completion event.
This patch doesn't remove delayed_work mechanism previously used,
instead it is only used as backup for when the standard approach fails.
Theoretically the delayed_work approach could also be removed, but it
was found to be too tricky to appopriately re-prime the evq if the event
wasn't found.