ON-16438: Handle efct_rxq flushes in efrm
Extend dma queue flush support to also support flushing efct_rxqs that
are attached to same evq as a txq.

Use a similar mechanism as the existing dma queue support where for each
efct_rxq attached to this vi/evq it will issue a flush, then when the
flush has completed efhw will call the flush handler. Also extends some
of the logic so that efct_rxqs are included when determining whether a
vi is expecting more flush completions.

Shared rxqs and rxqs which don't use the same evq as a txq will not use
this mechanism and continue using efct_rxq_release which both issues the
flush and releases the resource.